pandas Foundations and Manipulating DataFrames
The Rebrickable database includes data on every LEGO set that ever been
sold; the names of the sets, what bricks they contain, what color the
bricks are, etc. It might be small bricks, but this is big data! In this
project, we will get to explore the Rebrickable database. To do this we
need to know your way around pandas dataframes and pandas Foundations
and Manipulating DataFrames with pandas.