Data Manipulation ,Data Visualization ,Importing & Cleaning Data
There's a new era of data analysis in baseball. Using a new technology
called Statcast, Major League Baseball is now collecting the precise
location and movements of its baseballs and players. In this project, we
will use Statcast data to compare the home runs of two of baseball's
brightest (and largest) stars, Aaron Judge (6'7") and Giancarlo Stanton
(6'6"), both of whom now play for the New York Yankees.

To complete this project we should be comfortable with pandas and basic
data visualization in Python. Check out our Intermediate Python for Data
Science and Introduction to Data Visualization with Python courses to
get familiar or brush up on our skills.

The dataset used in this project is from Baseball
Savant["https://baseballsavant.mlb.com/about"]