DATA MANIPULATION,DATA VISUALIZATION & BOKEH
In 1854, Dr. John Snow (no, not the Game of Thrones's character) used a pre-computer method of spatial analysis by mapping patterns and occurrences of cholera outbreaks in Soho, London. He mapped the deaths in the neighbourhood and determined that a vast majority occurred around one particular water well and that those that died used that well. It is not only one of the earliest use of data visualization, but by solving this problem, he also founded spatial analysis and modern epidemiology.

In this Python project, we will reanalyze the data and recreate John Snow's famous map. This project is designed to test your knowledge of pandas and Bokeh, which can be acquired in the following courses: pandas Foundations and Interactive Data Visualization with Bokeh.